moved elastics system into it's own component
