+++ Assets/MRTK/SDK/Editor/Inspectors/UX/ObjectManipulator/ObjectManipulatorInspector.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-using Microsoft.MixedReality.Toolkit.Experimental.Physics;
 using Microsoft.MixedReality.Toolkit.UI;
 using Microsoft.MixedReality.Toolkit.Utilities;
 using Microsoft.MixedReality.Toolkit.Utilities.Editor;
@@ -41,23 +40,13 @@ namespace Microsoft.MixedReality.Toolkit.Editor
         private SerializedProperty onHoverEntered;
         private SerializedProperty onHoverExited;
 
-        private SerializedProperty elasticTypes;
-        private SerializedProperty translationElasticConfigurationObject;
-        private SerializedProperty rotationElasticConfigurationObject;
-        private SerializedProperty scaleElasticConfigurationObject;
-        private SerializedProperty translationElasticExtent;
-        private SerializedProperty rotationElasticExtent;
-        private SerializedProperty scaleElasticExtent;
+        private SerializedProperty elasticsManager;
 
         bool oneHandedFoldout = true;
         bool twoHandedFoldout = true;
         bool constraintsFoldout = true;
         bool physicsFoldout = true;
         bool smoothingFoldout = true;
-        bool elasticsFoldout = true;
-        bool translationElasticFoldout = false;
-        bool rotationElasticFoldout = false;
-        bool scaleElasticFoldout = false;
         bool eventsFoldout = true;
 
         public void OnEnable()
@@ -85,20 +74,14 @@ namespace Microsoft.MixedReality.Toolkit.Editor
             rotateLerpTime = serializedObject.FindProperty("rotateLerpTime");
             scaleLerpTime = serializedObject.FindProperty("scaleLerpTime");
 
+            // Elastics
+            elasticsManager = serializedObject.FindProperty("elasticsManager");
+
             // Manipulation Events
             onManipulationStarted = serializedObject.FindProperty("onManipulationStarted");
             onManipulationEnded = serializedObject.FindProperty("onManipulationEnded");
             onHoverEntered = serializedObject.FindProperty("onHoverEntered");
             onHoverExited = serializedObject.FindProperty("onHoverExited");
-
-            // Elastic configuration (ScriptableObject)
-            translationElasticConfigurationObject = serializedObject.FindProperty("translationElasticConfigurationObject");
-            rotationElasticConfigurationObject = serializedObject.FindProperty("rotationElasticConfigurationObject");
-            scaleElasticConfigurationObject = serializedObject.FindProperty("scaleElasticConfigurationObject");
-            translationElasticExtent = serializedObject.FindProperty("translationElasticExtent");
-            rotationElasticExtent = serializedObject.FindProperty("rotationElasticExtent");
-            scaleElasticExtent = serializedObject.FindProperty("scaleElasticExtent");
-            elasticTypes = serializedObject.FindProperty("elasticTypes");
         }
 
         public override void OnInspectorGUI()
@@ -178,45 +161,6 @@ namespace Microsoft.MixedReality.Toolkit.Editor
             }
 
             EditorGUILayout.Space();
-            elasticsFoldout = EditorGUILayout.Foldout(elasticsFoldout, "Elastics", true);
-
-            if (elasticsFoldout)
-            {
-                // This two-way enum cast is required because EnumFlagsField does not play nicely with
-                // SerializedProperties and custom enum flags.
-                var newElasticTypesValue = EditorGUILayout.EnumFlagsField("Manipulation types using elastic feedback: ", (TransformFlags)elasticTypes.intValue);
-                elasticTypes.intValue = (int)(TransformFlags)newElasticTypesValue;
-
-                // If the particular elastic type is requested, we offer the user the ability
-                // to configure the elastic system.
-                TransformFlags currentFlags = (TransformFlags)elasticTypes.intValue;
-
-                translationElasticFoldout = DrawElasticConfiguration<ElasticConfiguration>(
-                    "Translation Elastic",
-                    translationElasticFoldout,
-                    translationElasticConfigurationObject,
-                    translationElasticExtent,
-                    TransformFlags.Move,
-                    currentFlags);
-
-                rotationElasticFoldout = DrawElasticConfiguration<ElasticConfiguration>(
-                    "Rotation Elastic",
-                    rotationElasticFoldout,
-                    rotationElasticConfigurationObject,
-                    rotationElasticExtent,
-                    TransformFlags.Rotate,
-                    currentFlags);
-
-                scaleElasticFoldout = DrawElasticConfiguration<ElasticConfiguration>(
-                    "Scale Elastic",
-                    scaleElasticFoldout,
-                    scaleElasticConfigurationObject,
-                    scaleElasticExtent,
-                    TransformFlags.Scale,
-                    currentFlags);
-            }
-
-            EditorGUILayout.Space();
             eventsFoldout = EditorGUILayout.Foldout(eventsFoldout, "Manipulation Events", true);
 
             if (eventsFoldout)
@@ -227,37 +171,16 @@ namespace Microsoft.MixedReality.Toolkit.Editor
                 EditorGUILayout.PropertyField(onHoverExited);
             }
 
+            EditorGUILayout.Space();
+
+            Microsoft.MixedReality.Toolkit.Experimental.Inspectors.ElasticsManagerInspector.DrawElasticsManagerLink(elasticsManager, mh.gameObject);
+
+            EditorGUILayout.Space();
+
             // reset foldouts style
             style.fontStyle = previousStyle;
 
             serializedObject.ApplyModifiedProperties();
-        }
-
-        private bool DrawElasticConfiguration<T>(
-            string name,
-            bool expanded,
-            SerializedProperty elasticProperty,
-            SerializedProperty extentProperty,
-            TransformFlags requiredFlag,
-            TransformFlags providedFlags) where T : ElasticConfiguration
-        {
-            if (providedFlags.HasFlag(requiredFlag))
-            {
-                bool result = false;
-                using (new EditorGUI.IndentLevelScope())
-                {
-                    result = InspectorUIUtility.DrawScriptableFoldout<T>(
-                        elasticProperty,
-                        name,
-                        expanded);
-                    EditorGUILayout.PropertyField(extentProperty, includeChildren: true);
-                }
-                return result;
-            }
-            else
-            {
-                return false;
-            }
         }
     }
 }
