+++ Assets/MRTK/SDK/Experimental/Editor/Inspectors/Elastics.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b3469f70871e2144ab8c52142626f075
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
