+++ Assets/MRTK/SDK/Experimental/Editor/Inspectors/Elastics/ElasticsManagerInspector.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: baedc6d5ecac92c47af994226d486e67
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
