+++ Assets/MRTK/SDK/Experimental/Elastic/Scripts/ElasticsManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 49eebbfcc03829f4ebed61c2617da0fc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
