+++ Assets/MRTK/SDK/Features/Input/Handlers/ObjectManipulator.cs
@@ -261,6 +261,18 @@ namespace Microsoft.MixedReality.Toolkit.UI
             set => scaleLerpTime = value;
         }
 
+        [SerializeField]
+        [Tooltip("Elastics Manager slot to enable elastics simulation when manipulating the object.")]
+        private ElasticsManager elasticsManager;
+        /// <summary>
+        /// Elastics Manager slot to enable elastics simulation when manipulating the object.
+        /// </summary>
+        public ElasticsManager ElasticsManager
+        {
+            get => elasticsManager;
+            set => elasticsManager = value;
+        }
+
         #endregion Serialized Fields
 
         #region Event handlers
@@ -317,112 +329,13 @@ namespace Microsoft.MixedReality.Toolkit.UI
             set => onHoverExited = value;
         }
 
-        [Header("Elastic")]
-        [SerializeField]
-        [Tooltip("Reference to the ScriptableObject which holds the elastic system configuration for translation manipulation.")]
-        private ElasticConfiguration translationElasticConfigurationObject = null;
-
-        /// <summary>
-        /// Reference to the ScriptableObject which holds the elastic system configuration for translation manipulation.
-        /// </summary>
-        public ElasticConfiguration TranslationElasticConfigurationObject
-        {
-            get => translationElasticConfigurationObject;
-            set => translationElasticConfigurationObject = value;
-        }
-
-        [SerializeField]
-        [Tooltip("Reference to the ScriptableObject which holds the elastic system configuration for rotation manipulation.")]
-        private ElasticConfiguration rotationElasticConfigurationObject = null;
-
-        /// <summary>
-        /// Reference to the ScriptableObject which holds the elastic system configuration for rotation manipulation.
-        /// </summary>
-        public ElasticConfiguration RotationElasticConfigurationObject
-        {
-            get => rotationElasticConfigurationObject;
-            set => rotationElasticConfigurationObject = value;
-        }
-
-        [SerializeField]
-        [Tooltip("Reference to the ScriptableObject which holds the elastic system configuration for scale manipulation.")]
-        private ElasticConfiguration scaleElasticConfigurationObject = null;
-
-        /// <summary>
-        /// Reference to the ScriptableObject which holds the elastic system configuration for scale manipulation.
-        /// </summary>
-        public ElasticConfiguration ScaleElasticConfigurationObject
-        {
-            get => scaleElasticConfigurationObject;
-            set => scaleElasticConfigurationObject = value;
-        }
-
-        [SerializeField]
-        [Tooltip("Extent of the translation elastic.")]
-        private VolumeElasticExtent translationElasticExtent;
-
-        /// <summary>
-        /// Extent of the translation elastic.
-        /// </summary>
-        public VolumeElasticExtent TranslationElasticExtent
-        {
-            get => translationElasticExtent;
-            set => translationElasticExtent = value;
-        }
-
-        [SerializeField]
-        [Tooltip("Extent of the rotation elastic.")]
-        private QuaternionElasticExtent rotationElasticExtent;
-
-        /// <summary>
-        /// Extent of the rotation elastic.
-        /// </summary>
-        public QuaternionElasticExtent RotationElasticExtent
-        {
-            get => rotationElasticExtent;
-            set => rotationElasticExtent = value;
-        }
-
-        [SerializeField]
-        [Tooltip("Extent of the scale elastic.")]
-        private VolumeElasticExtent scaleElasticExtent;
-
-        /// <summary>
-        /// Extent of the scale elastic.
-        /// </summary>
-        public VolumeElasticExtent ScaleElasticExtent
-        {
-            get => scaleElasticExtent;
-            set => scaleElasticExtent = value;
-        }
-
-        [SerializeField]
-        [Tooltip("Indication of which manipulation types use elastic feedback.")]
-        private TransformFlags elasticTypes = 0; // Default to none enabled.
-
-        /// <summary>
-        /// Indication of which manipulation types use elastic feedback.
-        /// </summary>
-        public TransformFlags ElasticTypes
-        {
-            get => elasticTypes;
-            set => elasticTypes = value;
-        }
-        #endregion
+        #endregion Event Handlers
 
         #region Private Properties
 
         private ManipulationMoveLogic moveLogic;
         private TwoHandScaleLogic scaleLogic;
         private TwoHandRotateLogic rotateLogic;
-
-        private IElasticSystem<Vector3> translationElastic;
-        private IElasticSystem<Quaternion> rotationElastic;
-        private IElasticSystem<Vector3> scaleElastic;
-
-        // Magnitude of the velocity at which the elastic systems will
-        // cease being simulated (if enabled) and the object will stop updating/moving.
-        private const float elasticVelocityThreshold = 0.001f;
 
         /// <summary>
         /// Holds the pointer and the initial intersection point of the pointer ray
@@ -460,7 +373,7 @@ namespace Microsoft.MixedReality.Toolkit.UI
         private bool IsOneHandedManipulationEnabled => manipulationType.HasFlag(ManipulationHandFlags.OneHanded) && pointerIdToPointerMap.Count == 1;
         private bool IsTwoHandedManipulationEnabled => manipulationType.HasFlag(ManipulationHandFlags.TwoHanded) && pointerIdToPointerMap.Count > 1;
 
-        #endregion
+        #endregion Private Properties
 
         #region MonoBehaviour Functions
 
@@ -470,36 +383,18 @@ namespace Microsoft.MixedReality.Toolkit.UI
             rotateLogic = new TwoHandRotateLogic();
             scaleLogic = new TwoHandScaleLogic();
 
-            InitializeElastics();
+            if (elasticsManager)
+            {
+                elasticsManager.InitializeElastics(HostTransform);
+            }
         }
         private void Start()
         {
             rigidBody = HostTransform.GetComponent<Rigidbody>();
             constraints = gameObject.EnsureComponent<ConstraintManager>();
         }
-        private void Update()
-        {
-            // If the user is not actively interacting with the object,
-            // we let the elastic systems continue simulating, to allow
-            // the object to naturally come to rest.
-            if (!isManipulationStarted)
-            {
-                if (elasticTypes.HasFlag(TransformFlags.Move) && translationElastic != null && translationElastic.GetCurrentVelocity().magnitude > elasticVelocityThreshold)
-                {
-                    HostTransform.position = translationElastic.ComputeIteration(HostTransform.position, Time.deltaTime);
-                }
-                if (elasticTypes.HasFlag(TransformFlags.Rotate) && rotationElastic != null && rotationElastic.GetCurrentVelocity().eulerAngles.magnitude > elasticVelocityThreshold)
-                {
-                    HostTransform.rotation = rotationElastic.ComputeIteration(HostTransform.rotation, Time.deltaTime);
-                }
-                if (elasticTypes.HasFlag(TransformFlags.Scale) && scaleElastic != null && scaleElastic.GetCurrentVelocity().magnitude > elasticVelocityThreshold)
-                {
-                    HostTransform.localScale = scaleElastic.ComputeIteration(HostTransform.localScale, Time.deltaTime);
-                }
-            }
-        }
 
-        #endregion MonoBehaviour Functions
+        #endregion
 
         #region Private Methods
 
@@ -599,31 +494,6 @@ namespace Microsoft.MixedReality.Toolkit.UI
             return false;
         }
 
-        private void InitializeElastics()
-        {
-            if (elasticTypes.HasFlag(TransformFlags.Move))
-            {
-                translationElastic = new VolumeElasticSystem(HostTransform.position,
-                                                             translationElastic?.GetCurrentVelocity() ?? Vector3.zero,
-                                                             translationElasticExtent,
-                                                             translationElasticConfigurationObject.ElasticProperties);
-            }
-            if (elasticTypes.HasFlag(TransformFlags.Rotate))
-            {
-                rotationElastic = new QuaternionElasticSystem(HostTransform.rotation,
-                                                              rotationElastic?.GetCurrentVelocity() ?? Quaternion.identity,
-                                                              rotationElasticExtent,
-                                                              rotationElasticConfigurationObject.ElasticProperties);
-            }
-            if (elasticTypes.HasFlag(TransformFlags.Scale))
-            {
-                scaleElastic = new VolumeElasticSystem(HostTransform.localScale,
-                                                       scaleElastic?.GetCurrentVelocity() ?? Vector3.zero,
-                                                       scaleElasticExtent,
-                                                       scaleElasticConfigurationObject.ElasticProperties);
-            }
-        }
-
         #endregion Private Methods
 
         #region Public Methods
@@ -676,7 +546,10 @@ namespace Microsoft.MixedReality.Toolkit.UI
                     pointerIdToPointerMap.Add(id, new PointerData(eventData.Pointer, eventData.Pointer.Result.Details.Point));
 
                     // Re-initialize elastic systems.
-                    InitializeElastics();
+                    if (elasticsManager)
+                    {
+                        elasticsManager.InitializeElastics(HostTransform);
+                    }
 
                     // Call manipulation started handlers
                     if (IsTwoHandedManipulationEnabled)
@@ -882,6 +755,10 @@ namespace Microsoft.MixedReality.Toolkit.UI
             }
 
             constraints.Initialize(new MixedRealityTransform(HostTransform));
+            if (elasticsManager)
+            {
+                elasticsManager.EnableElasticsUpdate(false);
+            }
         }
 
         private void HandleManipulationEnded(Vector3 pointerGrabPoint, Vector3 pointerVelocity, Vector3 pointerAnglularVelocity)
@@ -902,6 +779,10 @@ namespace Microsoft.MixedReality.Toolkit.UI
             }
 
             ReleaseRigidBody(pointerVelocity, pointerAnglularVelocity);
+            if (elasticsManager)
+            {
+                elasticsManager.EnableElasticsUpdate(true);
+            }
         }
 
         #endregion Private Event Handlers
@@ -919,27 +800,21 @@ namespace Microsoft.MixedReality.Toolkit.UI
         {
             if (rigidBody == null)
             {
-                if (elasticTypes.HasFlag(TransformFlags.Move))
+                TransformFlags transformUpdated = 0;
+                if (elasticsManager)
                 {
-                    HostTransform.position = translationElastic.ComputeIteration(targetTransform.Position, Time.deltaTime);
+                    transformUpdated = elasticsManager.ApplyTargetTransform(targetTransform);
                 }
-                else
+
+                if (!transformUpdated.HasFlag(TransformFlags.Move))
                 {
                     HostTransform.position = isSmoothing ? Smoothing.SmoothTo(HostTransform.position, targetTransform.Position, moveLerpTime, Time.deltaTime) : targetTransform.Position;
                 }
-                if (elasticTypes.HasFlag(TransformFlags.Rotate))
-                {
-                    HostTransform.rotation = rotationElastic.ComputeIteration(targetTransform.Rotation, Time.deltaTime);
-                }
-                else
+                if (!transformUpdated.HasFlag(TransformFlags.Rotate))
                 {
                     HostTransform.rotation = isSmoothing ? Smoothing.SmoothTo(HostTransform.rotation, targetTransform.Rotation, rotateLerpTime, Time.deltaTime) : targetTransform.Rotation;
                 }
-                if (elasticTypes.HasFlag(TransformFlags.Scale))
-                {
-                    HostTransform.localScale = scaleElastic.ComputeIteration(targetTransform.Scale, Time.deltaTime);
-                }
-                else
+                if (!transformUpdated.HasFlag(TransformFlags.Scale))
                 {
                     HostTransform.localScale = isSmoothing ? Smoothing.SmoothTo(HostTransform.localScale, targetTransform.Scale, scaleLerpTime, Time.deltaTime) : targetTransform.Scale;
                 }
@@ -1076,3 +951,4 @@ namespace Microsoft.MixedReality.Toolkit.UI
         #endregion
     }
 }
+
